requirements.txt
Werkzeug==2.3.4
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-Cors==3.0.10
beautifulsoup4==4.12.2
python-dotenv==1.0.0
sumy==0.11.0
numpy==1.24.3
nltk==3.8.1